cloud/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bases==0.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
ecdsa==0.19.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
json5==0.12.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
morphys==1.0
multiformats==0.3.1.post4
multiformats-config==0.3.1
py-multibase==1.0.3
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
python-baseconv==1.2.2
python-dotenv==1.1.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
rsa==4.9.1
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tinydb==4.8.2
typing-inspection==0.4.0
typing-validation==1.2.12
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wheel==0.45.1
